+++ app/scripts/lib/lforms-util.js
@@ -815,8 +815,7 @@ LForms.Util = {
 
 
   /**
-   * We are transitioning lforms fields representing code (form.code, form.questionCode,
-   * items[x].questionCode
+   * We are transitioning lforms fields representing code (form.code, items[x].questionCode
    * and items[x].questionCodeSystem) to FHIR definition of Coding type.
    * In lforms, these fields are string type and FHIR Coding is an array of
    * objects encapsulating multiple codes
@@ -841,6 +840,7 @@ LForms.Util = {
     var display = isItem ? formOrItem.question : formOrItem.name;
     var codeSystemUrl = LForms.Util.getCodeSystem(codeSystem);
 
+    // if there is code
     if(code) {
       if(!formOrItem.codeList) {
         formOrItem.codeList = [];
@@ -848,7 +848,7 @@ LForms.Util = {
       var codeList = formOrItem.codeList;
       var found = false;
       for(var i = 0; i < codeList.length; i++) {
-        if(code === codeList[i].code && codeSystemUrl === codeList[i].system) {
+        if(code === codeList[i].code && (!codeSystemUrl && !codeList[i].system || codeSystemUrl === codeList[i].system)) {
           found = true;
           break;
         }
@@ -857,13 +857,17 @@ LForms.Util = {
       // if form data is converted from a FHIR Questionnaire that has no 'code' on items,
       // don't create a 'code' when converting it back to Questionnaire.
       if(!found && codeSystemUrl !== 'LinkId') {
-        codeList.unshift({
-          system: codeSystemUrl,
+        var code = {
           code: code,
           display: display
-        });
+        };
+        if (codeSystemUrl) {
+          code.system = codeSystemUrl;
+        }
+        codeList.unshift(code);
       }
     }
+    // if there is a codeList
     else {
       if(formOrItem.codeList && formOrItem.codeList.length > 0) {
         if(isItem) {
@@ -925,9 +929,6 @@ LForms.Util = {
     switch (codeSystemInLForms) {
       case "LOINC":
         codeSystem = "http://loinc.org";
-        break;
-      case undefined:
-        codeSystem = "http://unknown"; // temp solution. as code system is required for coding
         break;
       default:
         codeSystem = codeSystemInLForms;