+++ bower-dist/lforms.js
@@ -8182,14 +8182,7 @@ angular.module('lformsWidget').controller('LFormsCtrl', ['$window', '$scope', '$
 
 
   $scope.getTrustedCodingInstructions = function (item) {
-    var ret = '';
-    var instruction = item.codingInstructionsFormat === 'html' ? item.codingInstructionsXHTML ? item.codingInstructionsXHTML : item.codingInstructions : item.codingInstructions;
-
-    if (instruction) {
-      ret = $sce.trustAsHtml(instruction);
-    }
-
-    return ret;
+    return item.codingInstructions ? $sce.trustAsHtml(item.codingInstructions) : '';
   };
   /**
    * Get the sequence number for the current repeating item
@@ -11879,8 +11872,7 @@ LForms.Util = {
   },
 
   /**
-   * We are transitioning lforms fields representing code (form.code, form.questionCode,
-   * items[x].questionCode
+   * We are transitioning lforms fields representing code (form.code, items[x].questionCode
    * and items[x].questionCodeSystem) to FHIR definition of Coding type.
    * In lforms, these fields are string type and FHIR Coding is an array of
    * objects encapsulating multiple codes
@@ -11902,7 +11894,7 @@ LForms.Util = {
     var code = isItem ? formOrItem.questionCode : formOrItem.code;
     var codeSystem = isItem ? formOrItem.questionCodeSystem : formOrItem.codeSystem;
     var display = isItem ? formOrItem.question : formOrItem.name;
-    var codeSystemUrl = LForms.Util.getCodeSystem(codeSystem);
+    var codeSystemUrl = LForms.Util.getCodeSystem(codeSystem); // if there is code
 
     if (code) {
       if (!formOrItem.codeList) {
@@ -11913,7 +11905,7 @@ LForms.Util = {
       var found = false;
 
       for (var i = 0; i < codeList.length; i++) {
-        if (code === codeList[i].code && codeSystemUrl === codeList[i].system) {
+        if (code === codeList[i].code && (!codeSystemUrl && !codeList[i].system || codeSystemUrl === codeList[i].system)) {
           found = true;
           break;
         }
@@ -11922,24 +11914,30 @@ LForms.Util = {
 
 
       if (!found && codeSystemUrl !== 'LinkId') {
-        codeList.unshift({
-          system: codeSystemUrl,
+        var code = {
           code: code,
           display: display
-        });
+        };
+
+        if (codeSystemUrl) {
+          code.system = codeSystemUrl;
+        }
+
+        codeList.unshift(code);
       }
-    } else {
-      if (formOrItem.codeList && formOrItem.codeList.length > 0) {
-        if (isItem) {
-          // questionCode is required, so this shouldn't happen??
-          formOrItem.questionCode = formOrItem.codeList[0].code;
-          formOrItem.questionCodeSystem = formOrItem.codeList[0].system;
-        } else {
-          formOrItem.code = formOrItem.codeList[0].code;
-          formOrItem.codeSystem = formOrItem.codeList[0].system;
+    } // if there is a codeList
+    else {
+        if (formOrItem.codeList && formOrItem.codeList.length > 0) {
+          if (isItem) {
+            // questionCode is required, so this shouldn't happen??
+            formOrItem.questionCode = formOrItem.codeList[0].code;
+            formOrItem.questionCodeSystem = formOrItem.codeList[0].system;
+          } else {
+            formOrItem.code = formOrItem.codeList[0].code;
+            formOrItem.codeSystem = formOrItem.codeList[0].system;
+          }
         }
       }
-    }
 
     return formOrItem;
   },
@@ -11984,11 +11982,6 @@ LForms.Util = {
     switch (codeSystemInLForms) {
       case "LOINC":
         codeSystem = "http://loinc.org";
-        break;
-
-      case undefined:
-        codeSystem = "http://unknown"; // temp solution. as code system is required for coding
-
         break;
 
       default:
@@ -12730,7 +12723,6 @@ LForms.HL7 = function () {
      * Constructs an OBX5 for a list item (CNE/CWE)
      * @param itemVal a value for a list item
      * @param dataType the data type of the item (CNE or CWE)
-     * @param answerCS the answer code system
      * @return the OBX5 field string
      */
     _generateOBX5: function _generateOBX5(itemVal, dataType, answerCS) {
@@ -12741,6 +12733,7 @@ LForms.HL7 = function () {
         // For non-coded values, the text goes in OBX 5.9
         rtn = this.delimiters.component.repeat(8) + itemVal.text;
       } else {
+        var answerCS = !itemVal.codeSystem ? "" : itemVal.codeSystem === 'LOINC' || itemVal.codeSystem === _fhir_fhir_common__WEBPACK_IMPORTED_MODULE_0__["LOINC_URI"] ? this.LOINC_CS : itemVal.codeSystem;
         rtn = code + this.delimiters.component + itemVal.text + this.delimiters.component + answerCS;
       }
 
@@ -12814,8 +12807,6 @@ LForms.HL7 = function () {
               itemObxArray[6] = unitName + this.delimiters.component + unitName + this.delimiters.component + this.LOINC_CS;
             }
 
-            var answerCS = !item.answerCodeSystem || item.answerCodeSystem == 'LOINC' || item.answerCodeSystem == _fhir_fhir_common__WEBPACK_IMPORTED_MODULE_0__["LOINC_URI"] ? this.LOINC_CS : item.answerCodeSystem;
-
             for (var i = 0, len = vals.length; i < len; ++i) {
               var val = vals[i]; // OBX4 - sub id
 
@@ -12828,7 +12819,7 @@ LForms.HL7 = function () {
 
 
               if (item.dataType === 'CNE' || item.dataType === 'CWE') {
-                itemObxArray[5] = this._generateOBX5(val, item.dataType, answerCS);
+                itemObxArray[5] = this._generateOBX5(val, item.dataType);
               } else if (item.dataType === 'DT' || item.dataType === 'DTM') {
                 var dv = typeof val === 'string' ? LForms.Util.stringToDate(val) : val;
                 itemObxArray[5] = dv.toString(item.dataType === 'DT' ? this._DT_FMT : this._DTM_FMT);
@@ -13639,7 +13630,7 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
             if (item._obsLinkPeriodExt) {
               duration = item._obsLinkPeriodExt.valueDuration; // optional
 
-              itemCodeSystem = item.questionCodeSystem || _this2.codeSystem;
+              itemCodeSystem = item.questionCodeSystem;
               if (itemCodeSystem === 'LOINC') itemCodeSystem = serverFHIR.LOINC_URI;
               fhirjs = LForms.fhirContext.getFHIRAPI(); // a fhir.js client
 
@@ -14172,12 +14163,15 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
           itemId = 1; // for each item on this level
 
       for (var i = 0; i < iLen; i++) {
-        var item = items[i]; // question coding system. If form level code system is LOINC, assume all
-        // child items are of LOINC, unless specified otherwise.
-
-        if (this.type === "LOINC" && !item.questionCodeSystem) {
-          item.questionCodeSystem = "LOINC";
-        }
+        var item = items[i]; // item's code system is optional
+        // if (this.type ==="LOINC") {
+        //   if (!item.questionCodeSystem) {
+        //     item.questionCodeSystem = "LOINC";
+        //   }
+        //   if ((item.dataType === 'CNE' || item.dataType === 'CWE') && !item.answerCodeSystem) {
+        //     item.answerCodeSystem = "LOINC";
+        //   }
+        // }
 
         LForms.Util.initializeCodes(item); // set default dataType
 
@@ -14349,7 +14343,7 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
         } // add a link to external site for item's definition
 
 
-        if (item.questionCodeSystem === "LOINC") {
+        if (item.questionCodeSystem === "LOINC" || this.codeSystem === "LOINC" && !item.questionCodeSystem) {
           item._linkToDef = "http://s.details.loinc.org/LOINC/" + item.questionCode + ".html";
         } // process the sub items
 
@@ -14412,6 +14406,19 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
           "min": "0",
           "max": "1"
         };
+      }
+
+      if (!Array.isArray(item.answers) && item.answers !== "" && this.answerLists) {
+        item.answers = this.answerLists[item.answers];
+      } // answer code system
+
+
+      if (item.answerCodeSystem && Array.isArray(item.answers)) {
+        for (var i = 0, iLen = item.answers.length; i < iLen; i++) {
+          if (item.answers[i] && !item.answers[i].codeSystem) {
+            item.answers[i].codeSystem = item.answerCodeSystem;
+          }
+        }
       } // set up flags for question and answer cardinality
 
 
@@ -15674,12 +15681,8 @@ function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterat
           var answers = []; // 'answers' might be null even for CWE
           // need to recheck answers in case its value has been changed by data control
 
-          if (item.answers) {
-            if (angular.isArray(item.answers)) {
-              answers = item.answers;
-            } else if (item.answers !== "" && this.answerLists) {
-              answers = this.answerLists[item.answers];
-            }
+          if (Array.isArray(item.answers)) {
+            answers = item.answers;
           } // reset the modified answers (for the display text)